// old/data/type.ts
export type RUNE_TYPE =
	| "rune_orange"
	| "rune_purple"
	| "rune_green"
	| "rune_white"
	| "rune_black"
	| "rune_void"
	| "rune_blue"
	| "rune_red";

/**
 * TODO: 元素符文 枚举
 */
export enum RUNE_ENUM {
	/**
	 * TODO: 元素符文 - 虚无
	 */
	void = 'rune_void',
	/**
	 * TODO: 元素符文 - 诸海
	 */
	blue = 'rune_blue',
	/**
	 * TODO: 元素符文 - 烛火
	 */
	red = 'rune_red',
	/**
	 * TODO: 元素符文 - 界木
	 */
	green = 'rune_green',
	/**
	 * TODO: 元素符文 - 归忆
	 */
	orange = 'rune_orange',
	/**
	 * TODO: 元素符文 - 极雷
	 */
	purple = 'rune_purple',
	/**
	 * TODO: 元素符文 - 启程
	 */
	white = 'rune_white',
	/**
	 * TODO: 元素符文 - 焚绝
	 */
	black = 'rune_black'
};
// ...
export interface ROUTE_ANNEX_ARGS {
	/**
	 * TODO: 每个 时钟周期 中 传入 的 坐标
	 */
	location: server.Vector3;
	/**
	 * TODO: 每个 时钟周期 中 传入 的 维度
	 */
	dimension: server.Dimension;
	/**
	 * TODO: 每个 时钟周期 中 传入 的 时间积分
	 */
	tick: number;
};
// ...
export interface GET_PROPERTY_PANEL {
	/**
	 * TODO: 附着 - 元素符文
	 *
	 * @description 当实体被攻击时 被附加上的 元素符文
	 */
	add_rune: RUNE_TYPE;
	/**
	 * TODO: 自身 - 元素符文
	 *
	 * @description 当实体发动攻击时 附加到目标实体上的元素符文
	 */
	self_rune: RUNE_TYPE;
};
// ...
export interface ROUTE_ARGS {
	/**
	 * TODO: 坐标组
	 *
	 * @description 如果与< shoot >同时定义, 则 本参数失效
	 */
	locations: server.Vector3[];
	/**
	 * TODO: 运行维度
	 */
	dimension: server.Dimension;
	/**
	 * TODO: 在运行时 执行的代码
	 */
	on_move?: ROUTE_MOVE_CODE;
	/**
	 * TODO: 在结束时 执行的代码
	 */
	on_done?: ROUTE_CLOSE_CODE;
	/**
	 * TODO: 带有参数设定 的 粒子效果
	 */
	particleMolang?: [string, server.MolangVariableMap];
	/**
	 * TODO: 粒子效果
	 */
	particles?: string[];
	/**
	 * TODO: 冷却时间
	 */
	cooldown: number;
	/**
	 * TODO: 运行速度
	 */
	speed: number;
	/**
	 * TODO: 坐标偏移
	 */
	offset?: server.Vector3;
	/**
	 * TODO: 射击
	 *
	 * @description 如果与< locations >同时定义, 参数< locations >失效
	 */
	shoot?: {
		/**
		 * TODO: 射击起始位置
		 */
		start_place: server.Vector3;
		/**
		 * TODO: 射击方向
		 */
		toward: server.Vector3;
		/**
		 * TODO: 射击距离
		 */
		max_distance: number;
	};
};

/**
 * TODO: 路径执行计划表 运行时 的 回调函数
 *
 * @returns {boolean} 容器是否继续运行
 */
export interface ROUTE_MOVE_CODE { (data: ROUTE_ANNEX_ARGS): boolean; };

/**
 * TODO: 路径执行计划表 关闭时 的 回调函数
 */
export interface ROUTE_CLOSE_CODE { (data: ROUTE_ANNEX_ARGS): void; };